Add AOC day 7 part 2 solution
Even though part 1 and part 2 are largely similar, the parsing change of J from Jack to Joker means that they cannot use the same parser. Due to wanting to have part 1 and part 2 running simultaneously, this meant duplicating much of the underlying structs and enums to add the Joker. Perhaps with some additional effort this could be reduced, but currently Cards are being created using Card::try_into, so it is technically involved in parsing, and thus also needed to change.