Add AOC day 8 part 2 solution
I'm actually a bit surprised that the time to start the cycle and the cycle length were the same, thought I would have to figure out how to math that all together to see when the cycles syncronized up but they were nice. AOC isn't getting any coal this year from me!